chore: move services to folder by aws service
